feat: fill empty output from remote testing result
